+++ changelog.md
@@ -8,6 +8,8 @@ Added Tooltip docs.
 Fixed some crashing doc pages.
 
 CodeEditor:
+- Added symbol hover info and inline color manipulation through LX.ColorInput.
+- Support for `options.onHoverSymbol` to customize popup symbol data.
 - Suggestions icon, iconClass, sortText and filterText.
 - Added scope braces detection.